requirements.txt
asgiref==3.8.1
certifi==2024.12.14
cffi==1.17.1
charset-normalizer==3.4.1
cramjam==2.9.1
cryptography==44.0.0
defusedxml==0.8.0rc2
dj-database-url==2.3.0
Django==5.1.4
djangorestframework==3.15.2
djangorestframework_simplejwt==5.4.0
djoser==2.3.1
dnspython==2.7.0
gunicorn==23.0.0
idna==3.10
mongoengine==0.29.1
oauthlib==3.2.2
packaging==24.2
psycopg==3.2.3
psycopg-binary==3.2.3
pycparser==2.22
PyJWT==2.10.1
pymongo==4.10.1
python-dotenv==1.0.1
python-snappy==0.7.3
python3-openid==3.2.0
requests==2.32.3
requests-oauthlib==2.0.0
social-auth-app-django==5.4.2
social-auth-core==4.5.4
sqlparse==0.5.3
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.3.0
uuid==1.30
winkerberos==0.12.0
